requirements.txt
requests
discord.py
bs4
python-dotenv
lxml
argparse
selenium